Make terminal work in vscode and beyond
This has been a pain since this repo was created, but it turns out the
bug is related to antigen application. If we apply everytime, we ensure
oh-my-zsh is loaded correctly and themes work. While we're here, I also
sourced the nvm plugin from oh-my-zsh, and removed and removed avn.